% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corekit.R
\docType{package}
\name{corekit}
\alias{corekit}
\alias{corekit-package}
\title{corekit: a package to support the development of CORE-KIT.}
\description{
corekit: a package to support the development of CORE-KIT.
}
