% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources.R
\name{get_resources}
\alias{get_resources}
\title{Get resource estimates}
\usage{
get_resources(longlat, radius, path_to_wind, path_to_solar)
}
\arguments{
\item{longlat}{vector of length two, longitude and latitude.}

\item{radius}{radius of the buffer used to}

\item{path_to_wind}{path to a raster file which contains wind data}

\item{path_to_solar}{path to a raster file which contains solar data}
}
\description{
Function gets resource estimates from rasters, location and radius inputs
}
